Automated calculation of derived histograms and Bug fixes
We implemented a DerivedHistoIterator which is sufficient generic to handle the calculation of ALL derived histograms. It uses a method getDerivedHistoCalculator() to fetch  a calculator from each of the tasks used in the analysis of events. A few more checks to be done. It seems to work OK... and it handles the calculation seamlessly. Note: if the  number of bins and boundary are changed at run time (relative to default values), the changes will be automatically propagated to the derived histograms using the setParameters of the Configuration class.