Major change: Configuration from txt file
This new version of CAP features lots of changes...
(1) Configuration of the job is no longer done by a scritp but a txt file (by default AnalysisPart.txt). This file contains list of keyword, type, value which sets all the variables and parameters required for the execution of a task or analysis.
(2) The object model has been reduced: XXXXXDerivedCalculator have been removed and the creation and calculation of derived histograms has been (re)integrated into the main XXXXAnalyzer classes. Fewer classes, less to maintain...
(3) The naming convention of i/o files has been slightly modified. The XXXAnalyzer tasks' name is now the root of the file name on input/output for histograms of the corresponding type. When partial saves are performed, the files are saved into sub-directories  Partial_#/ where # is an index corresponding to the save index.
(4) The scale of B2 histogram is now divided by the rapidity bin width so that projections onto the x-axis behave like functions... (instead of histograms)